// src/types/api-types.ts
import { Bar, CartItemType, Line, Order, Pie, Product, ShippingInfo, Stats, User } from "./types";

export type MessageResponse = {
    success: boolean,
    message: string,
}

export type UserResponse = {
    success: boolean;
    user: User;
}

export type AllUserResponse = {
    success: boolean;
    users: User[];
}

export type AllProductResponse = {
    success: boolean;
    products: Product[];
}

export type CategoriesResponse = {
    success: boolean;
    categories: string[];
}

export type SearchProductsResponse = {
    success: boolean;
    products: Product[];
    totalPage: number;
}

export type StatesResponse = {
    success: boolean;
    stats: Stats;
}

export type PieResponse = {
    success: boolean;
    charts: Pie;
}

export type BarResponse = {
    success: boolean;
    charts: Bar;
}

export type LineResponse = {
    success: boolean;
    charts: Line;
}

export type SearchProductsRequest = {
    price: number;
    page: number;
    category: string;
    search: string;
    sort: string;
}

export type NewProductRequest = {
    id: string;
    formData: FormData;
}

export type UpdateProductRequest = {
    userId: string;
    productId: string;
    formData: FormData;
}

export type DeleteProductRequest = {
    userId: string;
    productId: string;
}

export type ProductResponse = {
    success: boolean,
    product: Product
}

export type NewOrderRequest = {
    shippingInfo: ShippingInfo;
    orderItems: CartItemType[];
    subtotal: number;
    tax: number;
    shippingCharges: number;
    discount: number;
    total: number;
    user: string;
}

export type UpdateOrderRequest = {
   userId: string;
   orderId: string;
}

export type CustomError = {
    status: number;
    data: {
        message: string;
        success: boolean;
    }
};

export type AllOrderResponse = {
    success: boolean;
    orders: Order[];
}

export type OrderDetailsResponse = {
    success: boolean;
    order: Order;
}

export type DeleteUserRequest = {
    userId: string;
    adminUserId: string;
}